6.1.3

+++ common/net/src/test/kotlin/calebxzhou/rdi/common/net/RHttpClientTest.kt
@@ -134,7 +134,7 @@ class RHttpClientTest {
             val body = RecordingBody("allowed".toByteArray(), "text/plain".toMediaType())
 
             assertEquals("allowed", request(body, *encodings.toTypedArray()).debugBodyForLogging())
-            assertEquals(1, body.writeCount, encodings)
+            assertEquals(1, body.writeCount.toDouble(), encodings)
         }
     }
